fix: build 에러 처리(이후에 다시 다 수정해야함)
